requirements.txt
streamlit==0.71.0
pandas==1.1.3
scikit-learn==0.23.2
numpy==1.19.2

imutils~=0.5.3
opencv-python~=4.2.0.34
pillow~=9.0.1
tensorflow~=1.15.0